Inventory
This class has a method for reading data from files and an overloaded version of that method to read the data from the file that the OrderWriter class writes. The data is written to parallel arrays and sorted by itemID with a bubble sort method. Inventory class also contains a random number generator to randomly select a prize from the prize array.